// src/utils/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful'
import { Document } from '@contentful/rich-text-types'

export interface IButtonFields {
  /** Label */
  label: string

  /** Href */
  href: string
}

export interface IButton extends Entry<IButtonFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'button'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IFooterFields {
  /** Name */
  name: string

  /** Social Links */
  socialLinks?: ISocialLink[] | undefined
}

export interface IFooter extends Entry<IFooterFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'footer'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IHeroBlockFields {
  /** Name */
  name?: string | undefined

  /** Hero Image */
  heroImage: Asset

  /** Button Label */
  buttonLabel?: string | undefined

  /** Button Href */
  buttonHref?: string | undefined

  /** Release Date Label */
  releaseDateLabel: string

  /** Subscribe Heading */
  subscribeHeading: string

  /** Subscribe Body */
  subscribeBody: string

  /** Content */
  content?: Document | undefined

  /** Subscribe Content */
  subscribeContent?: Document | undefined
}

export interface IHeroBlock extends Entry<IHeroBlockFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'HeroBlock'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IMetaFields {
  /** Title */
  title?: string | undefined

  /** Description */
  description?: string | undefined

  /** ISBN */
  isbn?: string | undefined

  /** Author */
  author?: string | undefined

  /** Release Date */
  release_date?: string | undefined
}

export interface IMeta extends Entry<IMetaFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'meta'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IPageFields {
  /** Name */
  name: string

  /** Title */
  title?: string | undefined

  /** Slug */
  slug: string

  /** Blocks */
  blocks?: (IHeroBlock | ISubscribeBlock)[] | undefined
}

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'page'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface ISocialLinkFields {
  /** Href */
  href: string

  /** image */
  image: Asset
}

export interface ISocialLink extends Entry<ISocialLinkFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'socialLink'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export type CONTENT_TYPE =
  | 'button'
  | 'footer'
  | 'HeroBlock'
  | 'meta'
  | 'page'
  | 'socialLink'

export type IEntry =
  | IButton
  | IFooter
  | IHeroBlock
  | IMeta
  | IPage
  | ISocialLink

export type LOCALE_CODE = 'en-US'

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-US'